M0018PQT-234 Use private conf `embedded` as the default for platform dependencies
This prevents an instance of "Too many loaded Platforms" error that
may occur when a platform depends on another platform.  It used to
occur when a platform depended on any zip because there was a bug in
the platform-loader.

By default only the default configuration is retrieved.  This commit
adds the embedded configuration.